lab 8 sanitizer work

+++ llvm/asan/asan_platform.ujos.c
@@ -96,7 +96,7 @@ static int
 asan_unpoison_shared_region(void *start, void *end, void *arg) {
     (void)start, (void)end, (void)arg;
     // LAB 8: Your code here
-    platform_asan_unpoison(start, (uintptr_t)end - (uintptr_t)start);
+    platform_asan_unpoison(start, (size_t)(end - start));
     return 0;
 }
